100-shell_sort.c

+++ 100-shell_sort.c
@@ -8,16 +8,16 @@
  */
 void shell_sort(int *array, size_t size)
 {
-	size_t i, j, h = 1;
+	long int i, j, h = 1;
 	int temp;
 
 	if (array == NULL || size == 0)
 		return;
-	while (h < size / 3)
+	while (h < (long int)(size / 3))
 		h = 3 * h + 1;
 	while (h > 0)
 	{
-		for (i = h; i < size; ++i)
+		for (i = h; i < (long int)size; ++i)
 		{
 			for (j = i; j >= h && array[j] < array[j - h]; j -= h)
 			{
